fix: remove and creates a unified schema
