--- src/app/schema/brand_product.js
@@ -1,8 +1,0 @@
-const BrandProductSchema = `
-  CREATE TABLE IF NOT EXISTS brand_product (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    name TEXT NOT NULL
-  );
-`;
-
-export default BrandProductSchema;
