--- src/app/schema/category_product.js
@@ -1,8 +1,0 @@
-const CategoryProductSchema = `
-  CREATE TABLE IF NOT EXISTS category_product (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    name TEXT NOT NULL
-  );
-`;
-
-export default CategoryProductSchema;
