--- src/app/schema/client_transactions.js
@@ -1,10 +1,0 @@
-const ClientTransactionsSchema = `  
-  CREATE TABLE IF NOT EXISTS client_transactions (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    uuid TEXT,    
-    user_uuid TEXT NULL,    
-    FOREIGN KEY(user_uuid) REFERENCES user(uuid)
-  );
-`;
-
-export default ClientTransactionsSchema;
