--- src/app/schema/color_product.js
@@ -1,8 +1,0 @@
-const ColorProductSchema = `
-  CREATE TABLE IF NOT EXISTS color_product (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    color TEXT NOT NULL
-  );
-`;
-
-export default ColorProductSchema;
