--- src/app/schema/gender_product.js
@@ -1,8 +1,0 @@
-const GenderProductSchema = `
-  CREATE TABLE IF NOT EXISTS gender_product (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    name TEXT NOT NULL
-  );
-`;
-
-export default GenderProductSchema;
