--- src/app/schema/revoked_tokens.ts
@@ -1,9 +1,0 @@
-const RevokedTokensSchema: string = `
-  CREATE TABLE RevokedTokens (
-    id SERIAL PRIMARY KEY,
-    token VARCHAR(255) UNIQUE NOT NULL,
-    revoked_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-  );
-`;
-
-export default RevokedTokensSchema;
