--- src/app/schema/size_product.js
@@ -1,8 +1,0 @@
-const SizeProductSchema = `
-  CREATE TABLE IF NOT EXISTS size_product (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    size TEXT NOT NULL
-  );
-`;
-
-export default SizeProductSchema;
