--- src/app/schema/user.js
@@ -1,16 +1,0 @@
-const UserSchema = `
-  CREATE TABLE IF NOT EXISTS user (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    uuid TEXT NOT NULL,
-    first_name TEXT NOT NULL,
-    last_name TEXT NOT NULL,
-    email TEXT NOT NULL,
-    password TEXT NOT NULL,
-    phone TEXT,
-    user_picture_name TEXT,
-    type TEXT DEFAULT 'user' NOT NULL,
-    UNIQUE (email)
-  );
-`;
-
-export default UserSchema;
